pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vmware-exit-agent"
version = "0.1.0"
description = "VMware Exit Intelligence Agent — Phase 1: deterministic, local, sample-driven analysis"
authors = [ { name = "Open Source" } ]
dependencies = [
    "pydantic>=1.8,<2.0",
    "PyYAML>=6.0",
    "rich>=11.0"
]
requires-python = ">=3.9"
